--- a/examples/undistort_image/src/main.rs
+++ b/examples/undistort_image/src/main.rs
@@ -66,7 +66,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // apply the remap
     let mut img_undistorted = Image::from_size_val(img.size(), 0.0)?;
     imgproc::interpolation::remap(
-        &img.clone().cast_and_scale(1.0 / 255.0)?,
+        &img.cast_and_scale(1.0 / 255.0)?,
         &mut img_undistorted,
         &map_x,
         &map_y,