… crates/kornia-tensor/src/storage.rs
… crates/kornia-tensor/src/tensor.rs